Fix de la gestion de reorderPlaylist et divers fix aux add current et public
Utilisation du playlistcontent_id plutôt que du kara_id et playlist_id

Ajout d'une fonction raisePosInPlaylist pour départager deux karas en
même position.

Closes #14
Fix des trucs dans #19 aussi.

+++ src/_common/db/userdata.sqlite3.sql
@@ -35,7 +35,7 @@ CREATE TABLE playlist_content (
     date_add       INTEGER,
     pseudo_add     TEXT,
     NORM_pseudo_add     TEXT COLLATE NOCASE,
-    pos            INTEGER NOT NULL,
+    pos            REAL NOT NULL,
     flag_playing   INTEGER NOT NULL,
     banned         INTEGER DEFAULT (0),
     ban_date       INTEGER,
